pyproject.toml
[project]
name = "asyncio-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial-asyncio>=0.6",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
]
